// frontend/src/client/types.gen.ts
export type AthleteCreate = {
    name: string;
    gender: string;
    id_number: string;
    kumite_category: string;
    kumite_individual: string;
    kumite_team: string;
    individual_kata: string;
    mixed_double_kata: string;
    team_kata: string;
    mixed_team_kata: string;
};